apps/model-importer/requirements.txt
biopython
biosimulators_cobrapy
biosimulators_utils >= 0.1.117
lxml
pyyaml
python_dateutil
requests
requests_cache
